Reorganize solomachine and cosmos components
