+++ crates/relayer-solomachine/src/impls/chain/ofa.rs
@@ -56,15 +56,15 @@ use ibc_relayer_types::timestamp::Timestamp;
 use ibc_relayer_types::tx_msg::Msg;
 use ibc_relayer_types::Height;
 
-use crate::impls::chain::components::channel_handshake_payload::BuildSolomachineChannelHandshakePayloads;
-use crate::impls::chain::components::connection_handshake_payload::BuildSolomachineConnectionHandshakePayloads;
-use crate::impls::chain::components::cosmos::query_client_state::QuerySolomachineClientStateFromCosmos;
-use crate::impls::chain::components::cosmos::query_consensus_state::QuerySolomachineConsensusStateFromCosmos;
-use crate::impls::chain::components::create_client_payload::BuildSolomachineCreateClientPayload;
-use crate::impls::chain::components::process_message::ProcessSolomachineMessages;
-use crate::impls::chain::components::receive_packet_payload::BuildSolomachineReceivePacketPayload;
-use crate::impls::chain::components::timeout_packet_payload::BuildSolomachineTimeoutPacketPayload;
-use crate::impls::chain::components::update_client_payload::BuildSolomachineUpdateClientPayload;
+use crate::impls::chain::cosmos_components::query_client_state::QuerySolomachineClientStateFromCosmos;
+use crate::impls::chain::cosmos_components::query_consensus_state::QuerySolomachineConsensusStateFromCosmos;
+use crate::impls::chain::solomachine_components::channel_handshake_payload::BuildSolomachineChannelHandshakePayloads;
+use crate::impls::chain::solomachine_components::connection_handshake_payload::BuildSolomachineConnectionHandshakePayloads;
+use crate::impls::chain::solomachine_components::create_client_payload::BuildSolomachineCreateClientPayload;
+use crate::impls::chain::solomachine_components::process_message::ProcessSolomachineMessages;
+use crate::impls::chain::solomachine_components::receive_packet_payload::BuildSolomachineReceivePacketPayload;
+use crate::impls::chain::solomachine_components::timeout_packet_payload::BuildSolomachineTimeoutPacketPayload;
+use crate::impls::chain::solomachine_components::update_client_payload::BuildSolomachineUpdateClientPayload;
 use crate::methods::encode::header::encode_header;
 use crate::methods::encode::sign_data::timestamped_sign_data_to_bytes;
 use crate::traits::solomachine::Solomachine;
