+++ crates/relayer-solomachine/src/impls/chain/solomachine_components/mod.rs
@@ -1,6 +1,5 @@
 pub mod channel_handshake_payload;
 pub mod connection_handshake_payload;
-pub mod cosmos;
 pub mod create_client_payload;
 pub mod process_message;
 pub mod query_client_state;
